+++ arcos4py/tools/__init__.py
@@ -5,10 +5,18 @@ __email__ = "benjamin.graedel@unibe.ch"
 __version__ = '0.2.1'
 
 from ._binarize_detrend import binData
-from ._cleandata import clipMeas, interpolation
-from ._detect_events import detectCollev, estimate_eps, track_events_dataframe, track_events_image
+from ._cleandata import clipMeas, interpolation, remove_image_background
+from ._detect_events import (
+    detectCollev,
+    estimate_eps,
+    track_events_dataframe,
+    track_events_image,
+    Linker,
+    ImageTracker,
+    DataFrameTracker,
+)
 from ._filter_events import filterCollev
-from ._stats import calcCollevStats
+from ._stats import calcCollevStats, calculate_statistics, calculate_statistics_per_frame
 
 __all__ = [
     "binData",
@@ -20,4 +28,10 @@ __all__ = [
     "estimate_eps",
     "track_events_dataframe",
     "track_events_image",
+    "calculate_statistics",
+    "calculate_statistics_per_frame",
+    "Linker",
+    "ImageTracker",
+    "DataFrameTracker",
+    "remove_image_background",
 ]